Implemented an adaptive intervention bucket class
The class intervention_bucket has been  created, inheriting from the Intervention superclass. The goal is to have an intervention package where different individual interventions can be triggered by individual conditions. The conditions will lie in an Event class that is yet to be implemented. The Sim class has been prepared for this.

As of now, the effects of interventions is read from two json files. This may need to be reviewed, and also the way they are imported.