Built events module
The events module has now been built and is imported by pathosim and events are integrated in the Sim class. Only a few events are built so far, and RatioEvent and RateEvent need more work before they are useable.

The interventions can now be run in an adaptive manner, responding to events. So far this is only integrated in the intervention_bucket class.